Separate out range.c from veneer.c and object.c
