+++ src/builtin/CMakeLists.txt
@@ -1,8 +1,8 @@
 target_sources(morpho
     PRIVATE
-        builtin.c
+        builtin.c builtin.h
         file.c
         functions.c
         system.c
-        veneer.c
+        veneer.c veneer.h
 )
