+++ src/classes/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(morpho
+    PRIVATE
+        range.c range.h
+)
