+++ src/datastructures/CMakeLists.txt
@@ -3,9 +3,9 @@ target_sources(morpho
         cmplx.c
         dictionary.c
         matrix.c
-        object.c
+        object.c object.h
         sparse.c
         syntaxtree.c
-        value.c
+        value.c value.h
         varray.c
 )
